requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.6
click==8.1.6
coloredlogs==15.0.1
dataclasses-json==0.5.14
elasticsearch==7.9.1
fastapi==0.99.1
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.6.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
langchain==0.0.266
langsmith==0.0.24
lxml==4.9.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.4.0
packaging==23.1
pinecone-io==0.0.1
posthog==3.0.1
protobuf==4.24.0
pulsar-client==3.2.0
pydantic==1.10.12
pydantic_core==2.4.0
PyPika==0.48.9
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.13.3
tqdm==4.66.1
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
